Extract `WidgetLayout` component
A layout-only container for things that might go into a widget.
