+++ static/app/views/dashboards/widgets/widgetFrame/widgetFrame.tsx
@@ -1,3 +1,4 @@
+import {Fragment} from 'react';
 import styled from '@emotion/styled';
 
 import Badge, {type BadgeProps} from 'sentry/components/badge/badge';
@@ -11,16 +12,12 @@ import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 
 import {ErrorPanel} from '../common/errorPanel';
-import {
-  MIN_HEIGHT,
-  MIN_WIDTH,
-  WIDGET_RENDER_ERROR_MESSAGE,
-  X_GUTTER,
-  Y_GUTTER,
-} from '../common/settings';
+import {WIDGET_RENDER_ERROR_MESSAGE} from '../common/settings';
 import {TooltipIconTrigger} from '../common/tooltipIconTrigger';
 import type {StateProps} from '../common/types';
 import {WarningsList} from '../common/warningsList';
+
+import {WidgetLayout} from './widgetLayout';
 
 export interface WidgetFrameProps extends StateProps {
   actions?: MenuItemProps[];
@@ -58,114 +55,113 @@ export function WidgetFrame(props: WidgetFrameProps) {
   const shouldShowActions = actions && actions.length > 0;
 
   return (
-    <Frame aria-label="Widget panel" borderless={props.borderless}>
-      <Header>
-        {props.warnings && props.warnings.length > 0 && (
-          <Tooltip title={<WarningsList warnings={props.warnings} />} isHoverable>
-            <TooltipIconTrigger aria-label={t('Widget warnings')}>
-              <IconWarning color="warningText" />
-            </TooltipIconTrigger>
-          </Tooltip>
-        )}
-
-        <Tooltip title={props.title} containerDisplayMode="grid" showOnlyOnOverflow>
-          <TitleText>{props.title}</TitleText>
-        </Tooltip>
-
-        {props.badgeProps &&
-          (Array.isArray(props.badgeProps) ? props.badgeProps : [props.badgeProps]).map(
-            (currentBadgeProps, i) => <RigidBadge key={i} {...currentBadgeProps} />
+    <WidgetLayout
+      ariaLabel="Widget panel"
+      Title={
+        <Fragment>
+          {props.warnings && props.warnings.length > 0 && (
+            <Tooltip title={<WarningsList warnings={props.warnings} />} isHoverable>
+              <TooltipIconTrigger aria-label={t('Widget warnings')}>
+                <IconWarning color="warningText" />
+              </TooltipIconTrigger>
+            </Tooltip>
           )}
 
-        {(props.description || shouldShowFullScreenViewButton || shouldShowActions) && (
-          <TitleHoverItems>
-            {props.description && (
-              // Ideally we'd use `QuestionTooltip` but we need to firstly paint the icon dark, give it 100% opacity, and remove hover behaviour.
-              <Tooltip
-                title={
-                  <span>
-                    {props.title && (
-                      <WidgetTooltipTitle>{props.title}</WidgetTooltipTitle>
-                    )}
-                    {props.description && (
-                      <WidgetTooltipDescription>
-                        {props.description}
-                      </WidgetTooltipDescription>
-                    )}
-                  </span>
-                }
-                containerDisplayMode="grid"
-                isHoverable
-              >
-                <WidgetTooltipButton
-                  aria-label={t('Widget description')}
-                  borderless
-                  size="xs"
-                  icon={<IconInfo size="sm" />}
-                />
-              </Tooltip>
+          <Tooltip title={props.title} containerDisplayMode="grid" showOnlyOnOverflow>
+            <TitleText>{props.title}</TitleText>
+          </Tooltip>
+
+          {props.badgeProps &&
+            (Array.isArray(props.badgeProps) ? props.badgeProps : [props.badgeProps]).map(
+              (currentBadgeProps, i) => <RigidBadge key={i} {...currentBadgeProps} />
             )}
-
-            {shouldShowActions && (
-              <TitleActionsWrapper
-                disabled={Boolean(props.actionsDisabled)}
-                disabledMessage={props.actionsMessage ?? ''}
-              >
-                {actions.length === 1 ? (
-                  actions[0]!.to ? (
-                    <LinkButton
-                      size="xs"
-                      disabled={props.actionsDisabled}
-                      onClick={actions[0]!.onAction}
-                      to={actions[0]!.to}
-                    >
-                      {actions[0]!.label}
-                    </LinkButton>
-                  ) : (
-                    <Button
-                      size="xs"
-                      disabled={props.actionsDisabled}
-                      onClick={actions[0]!.onAction}
-                    >
-                      {actions[0]!.label}
-                    </Button>
-                  )
-                ) : null}
-
-                {actions.length > 1 ? (
-                  <DropdownMenu
-                    items={actions}
-                    isDisabled={props.actionsDisabled}
-                    triggerProps={{
-                      'aria-label': t('Widget actions'),
-                      size: 'xs',
-                      borderless: true,
-                      showChevron: false,
-                      icon: <IconEllipsis direction="down" size="sm" />,
-                    }}
-                    position="bottom-end"
-                  />
-                ) : null}
-              </TitleActionsWrapper>
-            )}
-
-            {shouldShowFullScreenViewButton && (
-              <Button
-                aria-label={t('Open Full-Screen View')}
+        </Fragment>
+      }
+      Actions={
+        <Fragment>
+          {props.description && (
+            // Ideally we'd use `QuestionTooltip` but we need to firstly paint the icon dark, give it 100% opacity, and remove hover behaviour.
+            <Tooltip
+              title={
+                <span>
+                  {props.title && <WidgetTooltipTitle>{props.title}</WidgetTooltipTitle>}
+                  {props.description && (
+                    <WidgetTooltipDescription>
+                      {props.description}
+                    </WidgetTooltipDescription>
+                  )}
+                </span>
+              }
+              containerDisplayMode="grid"
+              isHoverable
+            >
+              <WidgetTooltipButton
+                aria-label={t('Widget description')}
                 borderless
                 size="xs"
-                icon={<IconExpand />}
-                onClick={() => {
-                  props.onFullScreenViewClick?.();
-                }}
+                icon={<IconInfo size="sm" />}
               />
-            )}
-          </TitleHoverItems>
-        )}
-      </Header>
+            </Tooltip>
+          )}
 
-      <VisualizationWrapper>
-        {props.error ? (
+          {shouldShowActions && (
+            <TitleActionsWrapper
+              disabled={Boolean(props.actionsDisabled)}
+              disabledMessage={props.actionsMessage ?? ''}
+            >
+              {actions.length === 1 ? (
+                actions[0]!.to ? (
+                  <LinkButton
+                    size="xs"
+                    disabled={props.actionsDisabled}
+                    onClick={actions[0]!.onAction}
+                    to={actions[0]!.to}
+                  >
+                    {actions[0]!.label}
+                  </LinkButton>
+                ) : (
+                  <Button
+                    size="xs"
+                    disabled={props.actionsDisabled}
+                    onClick={actions[0]!.onAction}
+                  >
+                    {actions[0]!.label}
+                  </Button>
+                )
+              ) : null}
+
+              {actions.length > 1 ? (
+                <DropdownMenu
+                  items={actions}
+                  isDisabled={props.actionsDisabled}
+                  triggerProps={{
+                    'aria-label': t('Widget actions'),
+                    size: 'xs',
+                    borderless: true,
+                    showChevron: false,
+                    icon: <IconEllipsis direction="down" size="sm" />,
+                  }}
+                  position="bottom-end"
+                />
+              ) : null}
+            </TitleActionsWrapper>
+          )}
+
+          {shouldShowFullScreenViewButton && (
+            <Button
+              aria-label={t('Open Full-Screen View')}
+              borderless
+              size="xs"
+              icon={<IconExpand />}
+              onClick={() => {
+                props.onFullScreenViewClick?.();
+              }}
+            />
+          )}
+        </Fragment>
+      }
+      Visualization={
+        props.error ? (
           <ErrorPanel error={error} />
         ) : (
           <ErrorBoundary
@@ -173,21 +169,11 @@ export function WidgetFrame(props: WidgetFrameProps) {
           >
             {props.children}
           </ErrorBoundary>
-        )}
-      </VisualizationWrapper>
-    </Frame>
+        )
+      }
+    />
   );
 }
-
-const TitleHoverItems = styled('div')`
-  display: flex;
-  align-items: center;
-  gap: ${space(0.5)};
-  margin-left: auto;
-
-  opacity: 1;
-  transition: opacity 0.1s;
-`;
 
 interface TitleActionsProps {
   children: React.ReactNode;
@@ -206,48 +192,6 @@ function TitleActionsWrapper({disabled, disabledMessage, children}: TitleActions
     </Tooltip>
   );
 }
-
-const Frame = styled('div')<{borderless?: boolean}>`
-  position: relative;
-  display: flex;
-  flex-direction: column;
-
-  height: 100%;
-  min-height: ${MIN_HEIGHT}px;
-  width: 100%;
-  min-width: ${MIN_WIDTH}px;
-
-  border-radius: ${p => p.theme.panelBorderRadius};
-  ${p => !p.borderless && `border: 1px solid ${p.theme.border};`}
-
-  background: ${p => p.theme.background};
-
-  :hover {
-    background-color: ${p => p.theme.surface200};
-    transition:
-      background-color 100ms linear,
-      box-shadow 100ms linear;
-    box-shadow: ${p => p.theme.dropShadowLight};
-  }
-
-  &:not(:hover):not(:focus-within) {
-    ${TitleHoverItems} {
-      opacity: 0;
-      ${p => p.theme.visuallyHidden}
-    }
-  }
-`;
-
-const HEADER_HEIGHT = '26px';
-
-const Header = styled('div')`
-  display: flex;
-  align-items: center;
-  height: calc(${HEADER_HEIGHT} + ${Y_GUTTER});
-  flex-shrink: 0;
-  gap: ${space(0.75)};
-  padding: ${X_GUTTER} ${Y_GUTTER} 0 ${X_GUTTER};
-`;
 
 const TitleText = styled(HeaderTitle)`
   ${p => p.theme.overflowEllipsis};
@@ -275,12 +219,4 @@ const WidgetTooltipButton = styled(Button)`
   pointer-events: none;
   padding-top: 0;
   padding-bottom: 0;
-`;
-
-const VisualizationWrapper = styled('div')`
-  display: flex;
-  flex-direction: column;
-  flex-grow: 1;
-  min-height: 0;
-  position: relative;
 `;
